requirements.txt
pyTelegramBotAPI==4.14.0
openai>=1.50.0
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
httpx>=0.24.0 